extract cached classes to their own files
