+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/Generic/CachedStrategySensitiveScorerBase.cs
@@ -3,7 +3,7 @@ using Raffinert.FuzzySharp.SimilarityRatio.Scorer.Generic;
 
 namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive.Generic;
 
-public abstract class StrategySensitiveScorerBase<T> : ScorerBase<T> where T : IEquatable<T>
+public abstract class CachedStrategySensitiveScorerBase<T> : CachedScorerBase<T> where T : IEquatable<T>
 {
-    protected abstract Func<T[], T[], int> Scorer { get; }
+    protected abstract Func<T[], int> Scorer { get; }
 }
