+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/TokenDifference/TokenDifferenceScorer.cs
@@ -1,8 +1,4 @@
-﻿using Raffinert.FuzzySharp.Extensions;
-using Raffinert.FuzzySharp.PreProcess;
-using Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive.Generic;
-using Raffinert.FuzzySharp.SimilarityRatio.Strategy;
-using Raffinert.FuzzySharp.SimilarityRatio.Strategy.Generic;
+﻿using Raffinert.FuzzySharp.SimilarityRatio.Strategy.Generic;
 using System;
 
 namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
@@ -10,30 +6,4 @@ namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
 public sealed class TokenDifferenceScorer : TokenDifferenceScorerBase
 {
     protected override Func<string[], string[], int> Scorer => DefaultRatioStrategy<string>.Calculate;
-}
-
-public sealed class CachedTokenDifferenceScorer : ICachedRatioScorer
-{
-    private readonly Func<string, string> _preprocessFactory;
-    private readonly CachedDefaultRatioStrategy<string> _scorer;
-
-    public CachedTokenDifferenceScorer(string input1, PreprocessMode preprocess = PreprocessMode.None)
-    {
-        _preprocessFactory = StringPreprocessorFactory.GetPreprocessor(preprocess);
-        input1 = _preprocessFactory(input1);
-        var tokens1 = input1.GetSortedWords();
-        _scorer = new CachedDefaultRatioStrategy<string>(tokens1);
-    }
-
-    public int Score(string input2)
-    {
-        input2 = _preprocessFactory(input2);
-        var tokens2 = input2.GetSortedWords();
-        return _scorer.Calculate(tokens2);
-    }
-
-    public void Dispose()
-    {
-        _scorer.Dispose();
-    }
 }
