+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/TokenDifference/TokenDifferenceScorerBase.cs
@@ -30,19 +30,3 @@ public abstract class TokenDifferenceScorerBase : StrategySensitiveScorerBase<st
         return Score(input1, input2);
     }
 }
-
-public abstract class CachedTokenDifferenceScorerBase : CachedStrategySensitiveScorerBase<string>, ICachedRatioScorer
-{
-    public override int Score(string[] input2)
-    {
-        return Scorer(input2);
-    }
-
-    public int Score(string input2)
-    {
-        var tokens2 = input2.GetSortedWords();
-
-        return Score(tokens2);
-    }
-    public abstract void Dispose();
-}
