+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/TokenSet/CachedTokenSetScorerBase.cs
@@ -1,24 +1,37 @@
-﻿using Raffinert.FuzzySharp.Extensions;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using Raffinert.FuzzySharp.Extensions;
 
 namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
 
-public abstract class TokenSetScorerBase : StrategySensitiveScorerBase
+public abstract class CachedTokenSetScorerBase : CachedScorerBase
 {
-    public override int Score(string input1, string input2)
+    public HashSet<string> Tokens1 { get; protected set; }
+    protected abstract FuzzySharp.Scorer Scorer { get; }
+
+    protected CachedTokenSetScorerBase()
     {
-        var tokens1 = new HashSet<string>(input1.SplitByAnySpace());
+        
+    }
+
+    protected CachedTokenSetScorerBase(string input1)
+    {
+        Tokens1 = new HashSet<string>(input1.SplitByAnySpace());
+    }
+
+    public override int Score(string input2)
+    {
         var tokens2 = new HashSet<string>(input2.SplitByAnySpace());
+        var tokens1 = new HashSet<string>(Tokens1);
 
         var intersection = GetIntersectionAndExcept(tokens1, tokens2);
 
         intersection.Sort();
-        
+
         var sortedIntersection = string.Join(" ", intersection);
-        var sortedDiff1To2     = (sortedIntersection + " " + string.Join(" ", tokens1.OrderBy(s => s))).Trim();
-        var sortedDiff2To1     = (sortedIntersection + " " + string.Join(" ", tokens2.OrderBy(s => s))).Trim();
+        var sortedDiff1To2 = (sortedIntersection + " " + string.Join(" ", tokens1.OrderBy(s => s))).Trim();
+        var sortedDiff2To1 = (sortedIntersection + " " + string.Join(" ", tokens2.OrderBy(s => s))).Trim();
 
         var score1 = Scorer(sortedIntersection, sortedDiff1To2);
         var score2 = Scorer(sortedIntersection, sortedDiff2To1);
