+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/TokenSet/TokenSetScorer.cs
@@ -6,20 +6,3 @@ public sealed class TokenSetScorer : TokenSetScorerBase
 {
     protected override FuzzySharp.Scorer Scorer => DefaultRatioStrategy.Calculate;
 }
-
-public sealed class CachedTokenSetScorer : CachedTokenSetScorerBase
-{
-    public CachedTokenSetScorer(string input1) : base(input1)
-    {
-    }
-
-    public CachedTokenSetScorer(CachedTokenSetScorerBase other)
-    {
-        Tokens1 = other.Tokens1;
-    }
-
-    protected override FuzzySharp.Scorer Scorer => DefaultRatioStrategy.Calculate;
-    public override void Dispose()
-    {   
-    }
-}
