+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/TokenSort/TokenSortScorer.cs
@@ -1,42 +1,8 @@
-﻿using Raffinert.FuzzySharp.Extensions;
-using Raffinert.FuzzySharp.SimilarityRatio.Strategy;
+﻿using Raffinert.FuzzySharp.SimilarityRatio.Strategy;
 
 namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
 
 public sealed class TokenSortScorer : TokenSortScorerBase
 {
     protected override FuzzySharp.Scorer Scorer => DefaultRatioStrategy.Calculate;
-}
-
-public sealed class CachedTokenSortScorer : CachedScorerBase
-{
-    private readonly ICachedStrategy _strategy;
-    private readonly bool _isStrategyOwner;
-
-    public CachedTokenSortScorer(string input1)
-    {
-        var sorted1 = input1.NormalizeSpacesAndSort();
-        _strategy = new CachedDefaultRatioStrategy(sorted1);
-        _isStrategyOwner = true;
-    }
-
-    public CachedTokenSortScorer(ICachedStrategy strategy, bool isStrategyOwner = false)
-    {
-        _strategy = strategy;
-        _isStrategyOwner = isStrategyOwner;
-    }
-
-    public override int Score(string input2)
-    {
-        var sorted2 = input2.NormalizeSpacesAndSort();
-        return _strategy.Calculate(sorted2);
-    }
-
-    public override void Dispose()
-    {
-        if(_isStrategyOwner)
-        {
-            _strategy.Dispose();
-        }
-    }
 }
